new subdomain definitions
